refactor(community): move PaginatedBoardList to widgets directory for better organization
- Updated imports in news and notice pages to reference PaginatedBoardList from the new widgets directory.
- Introduced PaginatedBoardList component to encapsulate pagination logic and improve modularity.
